Start on implementation of COERCE
COERCE is a multi that takes a value and a type object; users wishing
to define special semantics for how a value of a given type is coerced
to other type may implement the appropriate candidate for COERCE to
perform that coercion.  The default is call a method of the same name
as the type object's class on the value.