Ensure wrappers in Staging repos can be copied to their parent verbatim
We now generate wrappers that work in the parent repo when copied over.
They are not guaranteed to work in the Staging repo anymore though. When
the wrapper mode is `relative` and the staging and it's parent is in a
different dir, then the wrapper config would differ. Since we now create
wrappers in the staging repo with the parent repos config, those will not
work when located in the staging repo.